requirements.txt
PyGithub==1.58.1
requests==2.23.0
whatthepatch==0.0.6
